refactor: Separate enums and models into different directories
