+++ solvedac_community/Schemas/Enums/badge_category.py
