+++ solvedac_community/Schemas/Enums/badge_tier.py
