+++ solvedac_community/Schemas/Enums/class_decoration.py
