+++ solvedac_community/Schemas/Models/__init__.py
@@ -13,17 +13,16 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
-from dataclasses import dataclass
-from typing import Dict
-
-from .problem_level import ProblemLevel
-
-
-@dataclass
-class ProblemLevelData:
-    level: ProblemLevel
-    count: int
-
-    def __init__(self, data: Dict[str, int]):
-        self.level = ProblemLevel(data["level"])
-        self.count = data["count"]
+from .account_info import AccountInfo
+from .author import Author
+from .auto_completion_data import AutoCompletionData
+from .background import Background
+from .badge import Badge
+from .coinshop_product import CoinshopProduct
+from .organization import Organization
+from .problem_level_data import ProblemLevelData
+from .problem_search_data import ProblemSearchData
+from .problem_stats import ProblemStats
+from .solvedac_statistics import SolvedAcStatistics
+from .tagged_problem import TaggedProblem
+from .user import User
