+++ solvedac_community/Schemas/Models/problem.py
@@ -14,24 +14,17 @@ OTHER DEALINGS IN THE SOFTWARE.
 """
 
 from dataclasses import dataclass
-from typing import Dict, Union, List
+from typing import Dict, Union
 
-from .problem_level import ProblemLevel
-from .problem_tag import ProblemTag
+from solvedac_community.Schemas.Enums.problem_level import ProblemLevel
 
 
 @dataclass
-class TaggedProblem:
+class Problem:
     problem_id: int
     title_ko: str
     is_solvable: bool
     is_partial: bool
-    accepted_user_count: int
-    level: ProblemLevel
-    voted_user_count: int
-    is_level_locked: bool
-    average_tries: float
-    tags: List[ProblemTag]
 
     def __init__(self, data: Dict[str, Union[int, str, bool, list]]):
         self.problem_id: int = data["problemId"]
@@ -43,4 +36,3 @@ class TaggedProblem:
         self.voted_user_count: int = data["votedUserCount"]
         self.is_level_locked: bool = data["isLevelLocked"]
         self.average_tries: float = data["averageTries"]
-        self.tags: List[ProblemTag] = [ProblemTag(tag) for tag in data["tags"]]
