+++ solvedac_community/Schemas/Models/problem_stats.py
@@ -13,5 +13,23 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
-from . import Enums
-from . import Models
+from dataclasses import dataclass
+from typing import Dict
+
+from solvedac_community.Schemas.Enums.problem_level import ProblemLevel
+
+
+@dataclass
+class ProblemStats:
+    level: ProblemLevel
+    total: int
+    solved: int
+    tried: int
+    partial: int
+
+    def __init__(self, data: Dict[str, int]):
+        self.level = ProblemLevel(data["level"])
+        self.total = data["total"]
+        self.solved = data["solved"]
+        self.tried = data["tried"]
+        self.partial = data["partial"]
